Create dataStrategy.txt
First of all, what is Master Data Management (MDM)? Master data refers to the critical data that are essential to an enterprise’s business and often used in multiple disciplines. MDM is the establishment and maintenance of an enterprise-level data service that provides accurate, consistent and complete master data across the enterprise and to all business partners